+++ producer/api/src/receiver_discovery_service.cpp
@@ -100,7 +100,7 @@ ReceiversList ReceiverDiscoveryService::RotatedUriList(uint64_t nthread) {
     }
     ReceiversList list{uri_list_};
     lock.unlock();
-    auto shift = nthread % size;
+    auto shift = (int) nthread % size;
     std::rotate(list.begin(), list.begin() + shift, list.end());
     return list;
 }
